+++ lib/qc/src/semantic_validator_instructions.cc
@@ -15,6 +15,7 @@
 #include <quadrate/qc/ast_node_instruction.h>
 #include <quadrate/qc/ast_node_literal.h>
 #include <quadrate/qc/instructions.h>
+#include <quadrate/qc/numeric_literal.h>
 #include <quadrate/qc/semantic_validator.h>
 
 namespace Qd {